// src/generated-types/store-attribute.ts
export interface Id {
  id: string;
}

/** Message representing a status response */
export interface StatusResponse {
  success: boolean;
  message: string;
}

/** Request to get attributes by category ID */
export interface GetAttributesByCategoryIdRequest {
  categoryId: string;
}

/** Message representing a list of attributes */
export interface AttributeList {
  data: AttributeWithTranslations[];
}

/** Message representing an attribute */
export interface AttributeResponse {
  id: string;
  categoryId: string;
  slug: string;
  sortOrder: number;
}

/** Message representing an attribute with translations */
export interface AttributeWithTranslations {
  id: string;
  categoryId: string;
  slug: string;
  sortOrder: number;
  translations: AttributeTranslationResponse[];
}

/** Message representing an attribute translation */
export interface AttributeTranslationResponse {
  id: string;
  language: Language;
  name: string;
}